requirement.txt
asgiref==3.8.1
Django==5.0.4
django-cors-headers==4.3.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
pillow==10.3.0
PyJWT==2.8.0
pypng==0.20220715.0
python-dotenv==1.0.1
qrcode==7.4.2
sqlparse==0.4.4
typing_extensions==4.11.0
